Prevent Qt null ptr crash
For whatever reason, sometimes the QPostEventListLocker can be handed
an object which, while not null, contains a null private object.
This will cause a crash, so we check for it. Similarly, sometimes
the receiver object might be non-null but it's d_func() is, so
we check for that as well.

Why this shows up only on iOS is obscure, but it doesn't hurt to test
for it on other platforms.